+++ test/LightQuery.IntegrationTestsServer/Controllers/PaginatedLightQueryController.cs
@@ -15,6 +15,7 @@ namespace LightQuery.IntegrationTestsServer.Controllers
 
         private readonly LightQueryContext _context;
 
+        [HttpGet]
         [LightQuery(forcePagination: true, defaultPageSize: 3)]
         public IActionResult GetValues(bool returnEmptyList = false)
         {