+++ test/LightQuery.IntegrationTestsServer/Startup.cs
@@ -1,6 +1,9 @@
-﻿using Microsoft.AspNetCore.Builder;
+﻿using LightQuery.NSwag;
+using LightQuery.Swashbuckle;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.OpenApi.Models;
 
 namespace LightQuery.IntegrationTestsServer
 {
@@ -11,8 +14,36 @@ namespace LightQuery.IntegrationTestsServer
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddDbContext<LightQueryContext>(options => options.UseInMemoryDatabase(_inMemoryDatabaseName));
-            services.AddMvc();
 
+            services.AddSwaggerDocument(nSwagConfig =>
+            {
+                nSwagConfig.DocumentName = "swagger20";
+                nSwagConfig.OperationProcessors.Add(new LightQueryOperationsProcessor());
+            });
+            services.AddOpenApiDocument(nSwagConfig =>
+            {
+                nSwagConfig.DocumentName = "openapi30";
+                nSwagConfig.OperationProcessors.Add(new LightQueryOperationsProcessor());
+            });
+            
+            services.AddSwaggerGen(options =>
+            {
+                options.SwaggerDoc("swagger20", new OpenApiInfo()
+                {
+                    Description = "swagger20"
+                });
+                options.OperationFilter<LightQueryOperationFilter>();
+            });
+
+            services.AddSwaggerGen(options =>
+            {
+                options.SwaggerDoc("openapi30", new OpenApiInfo()
+                {
+                    Description = "openapi30"
+                });
+                options.OperationFilter<LightQueryOperationFilter>();
+            });
+            
 #if NETCORE3
             services.AddMvc(mvcOptions => mvcOptions.EnableEndpointRouting = false); ;
 #else
@@ -22,6 +53,32 @@ namespace LightQuery.IntegrationTestsServer
 
         public void Configure(IApplicationBuilder app)
         {
+            app.UseOpenApi(openApiConfig =>
+            {
+                openApiConfig.DocumentName = "swagger20";
+                openApiConfig.Path = "/nswag/swagger20.json";
+            });
+            app.UseOpenApi(openApiConfig =>
+            {
+                openApiConfig.DocumentName = "openapi30";
+                openApiConfig.Path = "/nswag/openapi30.json";
+            });
+            
+            /*
+             * Change the path for swageer json endpoints
+             * https://github.com/domaindrivendev/Swashbuckle.AspNetCore#change-the-path-for-swagger-json-endpoints
+             */
+            app.UseSwagger(options =>
+            {
+                options.RouteTemplate = "swashbuckle/{documentName}.json";
+                options.SerializeAsV2 = true;
+            });
+            
+            app.UseSwagger(options =>
+            {
+                options.RouteTemplate = "swashbuckle/{documentName}.json";
+            });
+
             app.UseMvc();
         }
     }